[mobile] add response for user one-time code
